requirements.txt
streamlit>=1.28,<2
python-dotenv>=1.0,<2
openai>=1.40,<2
